refactor: rename some structs and mods
